1/ refactoring

+++ app/src/main/java/com/example/check24tech/data/model/SaleItemDto.kt
@@ -4,6 +4,7 @@ import android.net.Uri
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.example.check24tech.utils.Constants
+import java.math.BigDecimal
 
 
 @Entity(tableName = Constants.TABLE_NAME)
@@ -12,7 +13,7 @@ data class SaleItemDto(
     var id: Int = 0,
     var title: String,
     var description: String? = null,
-    var price: Double? = null,
+    var price: BigDecimal? = null,
     var image: Uri? = null
 )
